refactor: Refactor asset and image classes, update namespaces
Renamed and reorganized asset-related classes: moved Asset.php and Cache.php to the root namespace, Image.php to Asset namespace, and Optimizer.php to Util as ImageOptimizer.php. Updated all relevant imports and usages throughout the codebase. Also renamed image responsive helper methods for clarity (e.g., buildSrcset to buildHtmlSrcset, getSizes to getHtmlSizes).
